requirements.txt
Flask==3.0.0
pytest==7.4.0
google-cloud-aiplatform==1.46.0
sentence-transformers==2.6.1